scripts/requirements.txt
pandas>=2.2.3
geopandas>=1.0.1
shapely>=2.0.6
scipy>=1.14.0
scikit-learn>=1.5.0
requests>=2.32.0
google-genai>=0.8.0
python-dotenv>=1.0.1
httpx>=0.27.0
tqdm>=4.66.5
numpy>=2.1.0
rasterio>=1.4.0
pyarrow>=17.0.0
openpyxl>=3.1.2
rapidfuzz>=3.10.0
